Move Task classes into Task package
